initial cut at moving Marlin to v0.10
I've replaced carrier with node-lstream due to various known issues, including
breakage in 0.10.  I've fixed a number of Node modules for the source
compatibility breakage and updated to newer ones that were already fixed.

At this point, I'm blocked on node issue 5112, which is that node doesn't
deliver header files in 0.10.  As a result, v8plus doesn't build with 0.10, and
both the "illumos_contract" and "zoneid" modules cannot build.